@@ -13,7 +13,7 @@ class ManagerController < ApplicationController
       @user.save!
     end
     flash[:notice] = t('info.moderator_panel.account_blocked')
-    ResqueMailer.delay.blocked(@user.id)
+    ResqueMailer.blocked(@user.id).deliver_later
     redirect_to :back
   end
 
