@@ -1,11 +1,11 @@
 #encoding: utf-8
 module ApplicationHelper
-  
+
   #ricarica i messaggi flash
   def reload_flash
     page.replace "flash_messages", partial: 'layouts/flash', locals: {flash: flash}
   end
-  
+
   def javascript(*args)
     content_for(:head) { javascript_include_tag(*args) }
   end
@@ -13,20 +13,20 @@ module ApplicationHelper
   def facebook_like
     "<div class=\"fb-like\" data-send=\"false\" data-layout=\"box_count\" data-width=\"100\" data-show-faces=\"false\"></div>"
   end
-  
+
   def calendar(*args)
     "<div id='calendar'></div>"
   end
 
 
- def resource_name
+  def resource_name
     :user
   end
- 
+
   def resource
     @resource ||= User.new
   end
- 
+
   def devise_mapping
     @devise_mapping ||= Devise.mappings[:user]
   end
@@ -42,10 +42,10 @@ module ApplicationHelper
           ret = I18n.l(from_time, format: :weekday) #this week
         end
       else
-        ret = I18n.l(from_time, format: :short)  #another week
+        ret = I18n.l(from_time, format: :short) #another week
       end
     elsif diff > 1.hours
-        ret = I18n.l(from_time, format: :hour)
+      ret = I18n.l(from_time, format: :hour)
     else
       ret = "<div data-countdown data-time='#{(from_time).to_i * 1000}' style='display:inline'></div>".html_safe
     end
@@ -57,6 +57,11 @@ module ApplicationHelper
     data = "otpauth://totp/#{user.email}?secret=#{user.rotp_secret}"
     url = "https://chart.googleapis.​com/chart?chs=200x200&chld=M|0&cht=qr&chl=#{data}"
     image_tag(url, alt: 'Google Authenticator QRCode')
+  end
+
+
+  def add_params(to_add={})
+    params.reject { |k, v| ['controller', 'action'].include? k }.merge to_add
   end
 
 end
