@@ -9,7 +9,7 @@ module Frm
       # If a user cannot be found, then no-op
       # This will happen if the user record has been deleted.
       if subscriber.present?
-        ResqueMailer.delay.topic_reply(post_id, subscriber.id)
+        ResqueMailer.topic_reply(post_id, subscriber.id).deliver_later
       end
     end
   end
