@@ -4,7 +4,7 @@ class CheckGroups
 
   def perform(*args)
     Group.joins(:participants).where({status: 'active', certified: false}).where(['groups.created_at < ?', Time.now - 7.days]).group('groups.id').having('count(users.*) < 2').readonly(false).each do |group|
-      ResqueMailer.delay.few_users_a(group.id)
+      ResqueMailer.few_users_a(group.id).deliver_later
       group.update_attribute(:status,Group::STATUS_FEW_USERS_A)
       group.update_attribute(:status_changed_at,Time.now)
     end
