@@ -9,7 +9,7 @@ Airesis::Application.configure do
   config.i18n.fallbacks = true
 
   # Disable Rails's static asset server (Apache or nginx will already do this)
-  config.serve_static_assets = false
+  config.serve_static_files = false
 
   # Compress JavaScripts and CSS
   config.assets.js_compressor = :uglifier
