@@ -1,7 +1,7 @@
 Airesis::Application.configure do
   config.cache_classes = true
   config.eager_load = false
-  config.serve_static_assets = true
+  config.serve_static_files = true
   config.static_cache_control = 'public, max-age=3600'
   config.consider_all_requests_local = false
   config.action_controller.perform_caching = false
@@ -10,6 +10,7 @@ Airesis::Application.configure do
   config.active_support.deprecation = :stderr
   #config.action_mailer.perform_deliveries = true
   config.action_mailer.delivery_method = :test
+  config.active_record.raise_in_transactional_callbacks = true
 end
 
 Airesis::Application.default_url_options = Airesis::Application.config.action_mailer.default_url_options
